+++ Entities/AHuman.h
@@ -877,6 +877,17 @@ ClassInfoGetters
 	/// <param name="newPrepTime">New duration to fully charge a throw in MS.</param>
 	void SetThrowPrepTime(long newPrepTime) { m_ThrowPrepTime = newPrepTime; }
 
+	/// <summary>
+	/// Gets this AHuman's stride sound. Ownership is NOT transferred!
+	/// </summary>
+	/// <returns>The SoundContainer for this AHuman's stride sound.</returns>
+	SoundContainer * GetStrideSound() const { return m_StrideSound; }
+
+	/// <summary>
+	/// Sets this AHuman's stride sound. Ownership IS transferred!
+	/// </summary>
+	/// <param name="newSound">The new SoundContainer for this AHuman's stride sound.</param>
+	void SetStrideSound(SoundContainer *newSound) { m_StrideSound = newSound; }
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Protected member variable and method declarations
@@ -928,7 +939,7 @@ protected:
     AtomGroup *m_pBGFootGroup;
     AtomGroup *m_BackupBGFootGroup;
     // The sound of the actor taking a step (think robot servo)
-    SoundContainer m_StrideSound;
+    SoundContainer *m_StrideSound;
     // Jetpack booster.
     AEmitter *m_pJetpack;
     // The max total time, in ms, that the jetpack can be used without pause